refactor(constants, main, otel): move OtlpProtocol type to a new module and update imports
