+++ src/common/core/constants.py
@@ -1,8 +1,7 @@
-from typing import Literal
+from common.core.types import OtlpProtocol
 
 DEFAULT_PROMETHEUS_MULTIPROC_DIR_NAME = "flagsmith-prometheus"
 
 LOGGING_DEFAULT_ROOT_LOG_LEVEL = "WARNING"
 
-OtlpProtocol = Literal["grpc", "http/protobuf"]
 DEFAULT_OTLP_PROTOCOL: OtlpProtocol = "http/protobuf"
