+++ src/common/core/types.py
@@ -1,0 +1,3 @@
+from typing import Literal
+
+OtlpProtocol = Literal["grpc", "http/protobuf"]
